+++ frontend/src/js/listeners/modal.js
@@ -1,0 +1,6 @@
+const openModal = function (nameFile, extension) {
+
+    
+}
+
+export { openModal }
