+++ index.html
@@ -102,7 +102,7 @@
     </div>
 
     <script type="module" src="frontend/src/js/init.js"></script>
-    <script type="module" src="frontend/src/js/modal.js"></script>
+    <script type="module" src="frontend/src/js/listeners.js"></script>
 
 </body>
 </html>
